Struct byteio::Reader

pub struct Reader<'a, R: ReadBytes<'a>>(_, _, _);

A convenience structure used for counting the number of bytes read.

This structure wraps an implementation of ReadBytes. It forwards read operations to the inner type while also maintaining a count of the number of bytes that pass through it.

When you have finished with it, you can return the original type via the into_inner method.

When compiled with the std feature this structure implements Read.

Examples

use byteio::{ReadBytes, ReadBytesExt, Reader};

fn main() -> byteio::Result<()> {
    let buf: &[u8] = &[1, 0, 2, 0, 0, 0, 3];

    let mut reader = Reader::new(buf);

    assert_eq!(reader.try_read_u8()?, 1);
    assert_eq!(reader.try_read_u16_be()?, 2);
    assert_eq!(reader.try_read_u32_be()?, 3);

    assert_eq!(reader.num_bytes_read(), 7);

    let inner = reader.into_inner();
    assert!(inner.is_empty());

    Ok(())
}

Implementations

impl<'a, R: ReadBytes<'a>> Reader<'a, R>

pub fn new(reader: R) -> Self

Creates a new Reader by wrapping a ReadBytes implementor.

Examples

use byteio::Reader;

let buf = [0_u8; 2];
let mut reader = Reader::new(&buf[..]);

pub fn num_bytes_read(&self) -> usize

Retrieves the number of bytes that have been read by this Reader.

Examples

use byteio::{ReadBytes, Reader};

let buf = [0_u8; 2];
let mut reader = Reader::new(&buf[..]);
let _ = reader.read_exact(1);
let _ = reader.read_exact(1);

assert_eq!(reader.num_bytes_read(), 2);

pub fn into_inner(self) -> R

Consumes this Reader and returns the original ReadBytes implementor.

Examples

use byteio::{ReadBytes, Reader};

let buf = [0_u8; 2];
let mut reader = Reader::new(&buf[..]);
let _ = reader.read_exact(1);
let inner = reader.into_inner();

assert_eq!(inner.len(), 1); // the reader consumed one byte from its view of the slice