Struct byteio::Writer

pub struct Writer<W: WriteBytes>(_, _);

A convenience structure used for counting the number of bytes written.

This structure wraps an implementation of WriteBytes. It forwards write operations to the inner type while also maintaining a count of the number of bytes that pass through it.

When you have finished with it, you can return the original type via the into_inner method.

When compiled with the std feature this structure implements Write.

Examples

use byteio::{WriteBytes, WriteBytesExt, Writer};

fn main() -> byteio::Result<()> {
    let mut buf = [0; 7];

    {
        let mut writer = Writer::new(&mut buf[..]);

        writer.try_write_u8(1)?;
        writer.try_write_u16_be(2)?;
        writer.try_write_u32_be(3)?;

        assert_eq!(writer.num_bytes_written(), 7);

        let inner = writer.into_inner();
        assert!(inner.is_empty());
    }

    assert_eq!(buf, [1, 0, 2, 0, 0, 0, 3]);

    Ok(())
}

Implementations

impl<W: WriteBytes> Writer<W>

pub fn new(writer: W) -> Self

Creates a new Writer by wrapping a WriteBytes implementor.

Examples

use byteio::Writer;

let mut buf = [0_u8; 2];
let mut writer = Writer::new(&mut buf[..]);

pub fn num_bytes_written(&self) -> usize

Retrieves the number of bytes that have been written by this Writer.

Examples

use byteio::{WriteBytes, Writer};

let mut buf = [0_u8; 2];
let mut writer = Writer::new(&mut buf[..]);
writer.write_exact(&[1]);
writer.write_exact(&[1]);

assert_eq!(writer.num_bytes_written(), 2);

pub fn into_inner(self) -> W

Consumes this Writer and returns the original WriteBytes implementor.

use byteio::{WriteBytes, Writer};

let mut buf = [0_u8; 2];
let mut writer = Writer::new(&mut buf[..]);
writer.write_exact(&[1]);
let inner = writer.into_inner();

assert_eq!(inner.len(), 1); // the writer consumed one byte from its view of the slice